+++ legacy/_1.11.2/src/main/kotlin/cn/enaium/noexpensive/callback/impl/AnvilTakeOutputCallbackImpl.kt
@@ -1,7 +1,6 @@
 package cn.enaium.noexpensive.callback.impl
 
 import cn.enaium.noexpensive.callback.AnvilTakeOutputCallback
-import net.minecraft.client.MinecraftClient
 import net.minecraft.client.resource.language.I18n
 import net.minecraft.entity.player.PlayerEntity
 import net.minecraft.item.ItemStack
@@ -9,9 +8,9 @@ import net.minecraft.item.ItemStack
 /**
  * @author Enaium
  */
-class AnvilTakeOutputCallbackImpl : AnvilTakeOutputCallback {
+abstract class AnvilTakeOutputCallbackImpl : AnvilTakeOutputCallback {
     override fun interact(output: ItemStack, player: PlayerEntity) {
-        MinecraftClient.getInstance().player == player || player.abilities.creativeMode && return
+        condition(player) && return
         if (!output.equals(ItemStack.EMPTY)) {
             val nbtCompound = output.getOrCreateNbtCompound("display")
             val list = nbtCompound.getList("Lore", 8)
@@ -24,4 +23,6 @@ class AnvilTakeOutputCallbackImpl : AnvilTakeOutputCallback {
             }
         }
     }
+
+    abstract fun condition(player: PlayerEntity): Boolean
 }