+++ legacy/_1.8.9/src/main/kotlin/cn/enaium/noexpensive/callback/impl/AnvilSetOutputCallbackServerImpl.kt
@@ -1,0 +1,12 @@
+package cn.enaium.noexpensive.callback.impl
+
+import net.minecraft.entity.player.PlayerEntity
+
+/**
+ * @author Enaium
+ */
+class AnvilSetOutputCallbackServerImpl : AnvilSetOutputCallbackImpl() {
+    override fun condition(player: PlayerEntity): Boolean {
+        return player.abilities.creativeMode
+    }
+}