+++ legacy/_1.8.9/src/main/kotlin/cn/enaium/noexpensive/callback/impl/AnvilTakeOutputCallbackImpl.kt
@@ -1,7 +1,6 @@
 package cn.enaium.noexpensive.callback.impl
 
 import cn.enaium.noexpensive.callback.AnvilTakeOutputCallback
-import net.minecraft.client.MinecraftClient
 import net.minecraft.client.resource.language.I18n
 import net.minecraft.entity.player.PlayerEntity
 import net.minecraft.item.ItemStack
@@ -10,9 +9,9 @@ import net.minecraft.item.ItemStack
 /**
  * @author Enaium
  */
-class AnvilTakeOutputCallbackImpl : AnvilTakeOutputCallback {
+abstract class AnvilTakeOutputCallbackImpl : AnvilTakeOutputCallback {
     override fun interact(output: ItemStack?, player: PlayerEntity) {
-        MinecraftClient.getInstance().player == player || player.abilities.creativeMode && return
+        condition(player) && return
         if (output != null) {
             val nbtCompound = output.getSubNbt("display", true)
             val list = nbtCompound.getList("Lore", 8)
@@ -25,4 +24,6 @@ class AnvilTakeOutputCallbackImpl : AnvilTakeOutputCallback {
             }
         }
     }
+
+    abstract fun condition(player: PlayerEntity): Boolean
 }