+++ modern/_1.17.1/src/main/kotlin/cn/enaium/noexpensive/NoExpensive.kt
@@ -3,9 +3,7 @@ package cn.enaium.noexpensive
 import cn.enaium.noexpensive.callback.AnvilSetOutputCallback
 import cn.enaium.noexpensive.callback.AnvilTakeOutputCallback
 import cn.enaium.noexpensive.callback.EnchantmentCanCombineCallback
-import cn.enaium.noexpensive.callback.impl.AnvilSetOutputCallbackImpl
-import cn.enaium.noexpensive.callback.impl.AnvilTakeOutputCallbackImpl
-import cn.enaium.noexpensive.callback.impl.EnchantmentCanCombineCallbackImpl
+import cn.enaium.noexpensive.callback.impl.*
 import cn.enaium.noexpensive.command.combineHigherCommand
 import cn.enaium.noexpensive.command.compatibilityCommand
 import cn.enaium.noexpensive.command.maxLevelCommand
@@ -29,10 +27,24 @@ fun initializer() {
     EnchantmentCanCombineCallback.EVENT.register(
         EnchantmentCanCombineCallbackImpl()
     )
-    AnvilSetOutputCallback.EVENT.register(AnvilSetOutputCallbackImpl())
-    AnvilTakeOutputCallback.EVENT.register(AnvilTakeOutputCallbackImpl())
     Config.load()
     Runtime.getRuntime().addShutdownHook(Thread { Config.save() })
+}
+
+object Client {
+    @JvmStatic
+    fun client() {
+        AnvilSetOutputCallback.EVENT.register(AnvilSetOutputCallbackClientImpl())
+        AnvilTakeOutputCallback.EVENT.register(AnvilTakeOutputCallbackClientImpl())
+    }
+}
+
+object Server {
+    @JvmStatic
+    fun server() {
+        AnvilSetOutputCallback.EVENT.register(AnvilSetOutputCallbackServerImpl())
+        AnvilTakeOutputCallback.EVENT.register(AnvilTakeOutputCallbackServerImpl())
+    }
 }
 
 val ROOT = CommandManager.literal("noexpensive")