+++ modern/_1.19.4/src/main/kotlin/cn/enaium/noexpensive/callback/impl/AnvilTakeOutputCallbackServerImpl.kt
@@ -1,0 +1,12 @@
+package cn.enaium.noexpensive.callback.impl
+
+import net.minecraft.entity.player.PlayerEntity
+
+/**
+ * @author Enaium
+ */
+class AnvilTakeOutputCallbackServerImpl : AnvilTakeOutputCallbackImpl() {
+    override fun condition(player: PlayerEntity): Boolean {
+        return player.abilities.creativeMode
+    }
+}