+++ modern/_1.20.6/src/main/kotlin/cn/enaium/noexpensive/callback/impl/AnvilTakeOutputCallbackImpl.kt
@@ -1,7 +1,6 @@
 package cn.enaium.noexpensive.callback.impl
 
 import cn.enaium.noexpensive.callback.AnvilTakeOutputCallback
-import net.minecraft.client.MinecraftClient
 import net.minecraft.component.DataComponentTypes
 import net.minecraft.component.type.LoreComponent
 import net.minecraft.entity.player.PlayerEntity
@@ -10,9 +9,9 @@ import net.minecraft.item.ItemStack
 /**
  * @author Enaium
  */
-class AnvilTakeOutputCallbackImpl : AnvilTakeOutputCallback {
+abstract class AnvilTakeOutputCallbackImpl : AnvilTakeOutputCallback {
     override fun interact(output: ItemStack, player: PlayerEntity) {
-        MinecraftClient.getInstance().player == player || player.abilities.creativeMode && return
+        condition(player) && return
         val lore = output[DataComponentTypes.LORE]
         lore ?: return
 
@@ -20,4 +19,6 @@ class AnvilTakeOutputCallbackImpl : AnvilTakeOutputCallback {
             LoreComponent(lore.lines.filterNot { it.toString().contains("container.repair.cost") })
         }
     }
+
+    abstract fun condition(player: PlayerEntity): Boolean
 }